app/requirements.txt
anyio==3.4.0
asgiref==3.4.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
dependency-injector==4.37.0
Deprecated==1.2.13
docker==5.0.3
fakeredis==1.7.0
fastapi==0.70.1
greenlet==1.1.2
h11==0.12.0
hiredis==2.0.0
httpcore==0.14.4
httptools==0.3.0
httpx==0.21.3
idna==3.3
iniconfig==1.1.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mirakuru==2.4.1
packaging==21.3
pluggy==1.0.0
port-for==0.6.1
psutil==5.9.0
py==1.11.0
pycparser==2.21
pydantic==1.9.0
PyNaCl==1.5.0
pyparsing==3.0.6
pytest==6.2.5
pytest-mockredis==0.1.6
python-dotenv==0.19.2
PyYAML==6.0
redis==4.0.2
redislite==1.0.254
requests==2.27.1
responses==0.16.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.29
starlette==0.16.0
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
uvloop==0.16.0
watchgod==0.7
websocket-client==1.2.3
websockets==10.1
wrapt==1.13.3
